third_party/blink/web_tests/fast/forms/range/range-set-attribute-expected.txt

Test to see if setting the value attribute updates the value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.setAttribute("value", "10"); input.value is "10"
rewriting the value attribute should update the value
PASS input.setAttribute("value", "20"); input.value is "20"
changing the max should effect value
PASS input.setAttribute("max", "10"); input.value is "10"
value attribute should not change the value after you set a value
PASS input.value = 10; input.setAttribute("value", "5"); input.value is "10"
Updating step attribute should adjust the last value.
PASS input.value is "50"
PASS input.step = "15"; input.value is "45"
PASS input.step = "20"; input.value is "40"
PASS successfullyParsed is true

TEST COMPLETE
